RENAME type file to match category class name
