+++ tests/test_somfy_api.py
@@ -2,7 +2,7 @@ import os
 
 import httpretty
 
-from src.api.devices.types import Category
+from src.api.devices.category import Category
 from src.api.somfy_api import BASE_URL, SomfyApi
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
